requirements.txt
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
numpy==1.23.4
pandas==1.5.1
Pillow==9.2.0
pytorch-lightning==1.7.7
torch==1.12.1
torchmetrics==0.10.1
torchvision==0.13.1
tqdm==4.64.1
typer==0.6.1
